printk 32 bit must be  carefulgit add *git add *git add *git add *,and can printk e820 map

+++ include/linux/types.h
@@ -18,8 +18,8 @@ typedef long long __s64;
 
 typedef unsigned char u8;
 typedef unsigned short u16;
-typedef unsigned long u32;
-typedef unsigned long long u64;
+typedef unsigned int u32;
+typedef unsigned long long  u64;
